Adjust Python3 code
I'm not sure that I made the right assumptions about Python3's Unicode
handling when I made the first patch to it.

By constructing the specific `diffs` output I created a sequence of code
units that `diff_main` in Python3 would _not_ have made because it's
operating on Unicode code points natively when finding the common
prefix.

Therefore I do not think that the Python3 library experienced this
problem as the others did. Nonetheless it _has_ been reporting the diff
length differently than in other languages and I have left that change
in there.

Of note, it doesn't look like we have true harmony between the languages
despite the appearance of such. The `lua` wiki page makes this clear,
but at least with Python we have the ability to harmonize the meaning of
the lengths and I have done that in this change.