Handle cases where we delete characters
In the previous iteration of this patch we were only properly handling
cases where a new surrogate pair was inserted in between two existing
pairs whose high surrogates all matched.

Unfortunately when swapping characters or performing any edits where we
delete a surrogate pair the patch failed because it only carried the
trailing high surrogate over to the next group instead of distributing
it to any insert _and_ delete groups following an equality group.

In this patch I've updated the JavaScript library to properly distribute
the trailing high surrogate.