Python: Update patch and handle wide/narrow compilation
Python can be compiled in "narrow" mode or "wide" mode which determines
how wide the internal string units are. In narrow mode we have UCS-2
code units and higher-order Unicode code points will be stored as
surrogate pairs. In wide mode we have UCS-4 code units and so all
Unicode code points will be stored as a single item in the string.

This patch incorporate a decision based on the internal string width to
run the native-looking `toDelta` or the _encoded_ version. In the
_encoded_ version we explicitly encode the string to `utf-16be` for
consistency sake with the other libraries.

We _could_ always run the encoded version but I suspect that the native
version will be faster and many deployments will be running narrow mode.